Fixing a major error
I am the biggest moron on the face of the planet...the Randomizer has not "removed" the tomatoes from Dropsanity, as a seed I just now played (which had one be Way of the Hoard) confirmed. Instead, the Plandomizer has them labeled as "Apple Gauntlet" enemies! Thanks a lot, @2Dos! X_x